Started work on a config.set rewrap function
I might fork secureconfig from bitbucket and put this in the config class as something like setsafe.
Currently, Goodshare is having a problem with saving the returned Goodreads tokens. I can't save the resulting gc.session.access_token and gc.session.access_token_secret directly into config, as it yells at me about having values in byte form or something.
So I ascii encode them with the included encode function, save them as access_token and access_token_secret, and put those values into config.
Unfortunately, the checks I put in to make sure that the resulting encoded tokens and the decrypted config values are the same keep failing for some reason -- I'm not sure if this has to do with it's encoding, though I dout it.
The setsafe function is intended to recheck the base variable against it's corresponding decrypted config value a few times and reset that config value.
I'm not sure if this will be affective at solving my particular problem, and i'm also not aware of the relation between the original value vs. It's decrypted config counterpart.